[FIX] use post_load website_sale_search_custom
It allows to avoid monkey pathes in databases where this module is not installed
