+++ website_sale_search_custom/README.rst
@@ -5,9 +5,4 @@ Custom website search
 * clear the search if the person clicks on any category link so the results are not filtered anymore.
 * adds results from searching by tags
 
-WARNING
--------
-
-Due to odoo restriction, this module should be installed on all databases in odoo instance.
-
 Tested on Odoo 8.0 f8d5a6727d3e8d428d9bef93da7ba6b11f344284
