+++ website_sale_search_custom/__init__.py
@@ -1,1 +1,4 @@
 import models
+
+def post_load():
+    import query_url
