+++ website_sale_search_custom/__openerp__.py
@@ -13,5 +13,6 @@
     'data': [
         'views.xml',
         ],
+    "post_load": 'post_load',
     'installable': True
 }
