+++ website_sale_search_custom/query_url.py
@@ -1,0 +1,30 @@
+import werkzeug
+
+from openerp.addons.website_sale.controllers import main as main_file
+
+
+class QueryURL(object):
+    def __init__(self, path='', **args):
+        self.path = path
+        self.args = args
+
+    def __call__(self, path=None, **kw):
+        if not path:
+            path = self.path
+        is_category = path.startswith('/shop/category/')
+        for k,v in self.args.items():
+            if is_category and k=='search':
+                continue
+            kw.setdefault(k,v)
+        l = []
+        for k,v in kw.items():
+            if v:
+                if isinstance(v, list) or isinstance(v, set):
+                    l.append(werkzeug.url_encode([(k,i) for i in v]))
+                else:
+                    l.append(werkzeug.url_encode([(k,v)]))
+        if l:
+            path += '?' + '&'.join(l)
+        return path
+
+main_file.QueryURL = QueryURL
